move linux-specific code to its own module
